+++ js/views/dataset.vue
@@ -111,6 +111,19 @@ export default {
                     icon: 'send',
                     method: this.transfer_request
                 });
+                if (!this.dataset.archived) {
+                actions.push({
+                    label: this._('Archive'),
+                    icon: 'archive',
+                    method: this.archive
+                })
+                } else {
+                actions.push({
+                    label: this._('Unarchive'),
+                    icon: 'undo',
+                    method: this.unarchive
+                });
+                }
                 if(!this.dataset.deleted) {
                     actions.push({
                         label: this._('Delete'),
@@ -169,6 +182,22 @@ export default {
         edit() {
             this.$go({name: 'dataset-edit', params: {oid: this.dataset.id}});
         },
+        archive() {
+            this.dataset.archived = new Date().toISOString();
+            API.datasets.update_dataset({dataset: this.dataset.id, payload: this.dataset},
+                (response) => {
+                    this.dataset.on_fetched(response);
+                }
+            );
+        },
+        unarchive() {
+            this.dataset.archived = null;
+            API.datasets.update_dataset({dataset: this.dataset.id, payload: this.dataset},
+                (response) => {
+                    this.dataset.on_fetched(response);
+                }
+            );
+        },
         confirm_delete() {
             this.$root.$modal(
                 require('components/dataset/delete-modal.vue'),
@@ -201,7 +230,7 @@ export default {
                     class: _class,
                     label
                 });
-            } else if (existing) {
+            } else if (!value && existing) {
                 this.badges.splice(this.badges.indexOf(existing), 1);
             }
         }