+++ js/views/reuse.vue
@@ -240,6 +240,18 @@ export default {
       this.$root.$modal(require('components/badges/modal.vue'), {
         subject: this.reuse
       })
+    },
+    addOrRemoveBadge(id, value, _class, label) {
+        const existing = this.badges.find(b => b.id === id);
+        if (value && !existing) {
+            this.badges.push({
+                id,
+                class: _class,
+                label
+            });
+        } else if (!value && existing) {
+            this.badges.splice(this.badges.indexOf(existing), 1);
+        }
     }
   },
   route: {
@@ -258,16 +270,10 @@ export default {
       }
     },
     'reuse.deleted': function (deleted) {
-      if (deleted) {
-        this.badges = [
-          {
-            class: 'danger',
-            label: this._('Deleted')
-          }
-        ]
-      } else {
-        this.badges = []
-      }
+      this.addOrRemoveBadge('deleted', deleted, 'danger', this._('Deleted'));
+    },
+    'reuse.archived': function(archived) {
+      this.addOrRemoveBadge('archived', archived, 'warning', this._('Archived'));
     }
   }
 }