+++ udata/core/owned.py
@@ -24,6 +24,22 @@ class OwnedQuerySet(UDataQuerySet):
             qs |= Q(owner=owner) | Q(organization=owner)
         return self(qs)
 
+    def visible_by_user(self, user: User, visible_query: Q):
+        """Return EVERYTHING visible to the user."""
+        if user.is_anonymous:
+            return self(visible_query)
+
+        if user.sysadmin:
+            return self()
+
+        owners: list[User | Organization] = list(user.organizations) + [user.id]
+        # We create a new queryset because we want a pristine self._query_obj.
+        owned_qs: OwnedQuerySet = self.__class__(self._document, self._collection_obj).owned_by(
+            *owners
+        )
+
+        return self(visible_query | owned_qs._query_obj)
+
 
 def check_owner_is_current_user(owner):
     from udata.auth import admin_permission, current_user