+++ udata/tests/organization/test_organization_rdf.py
@@ -4,6 +4,8 @@ from rdflib.namespace import FOAF, RDF, RDFS
 from rdflib.resource import Resource as RdfResource
 
 from udata import api
+from udata.core.dataservices.factories import DataserviceFactory
+from udata.core.dataservices.models import Dataservice
 from udata.core.dataset.factories import DatasetFactory
 from udata.core.dataset.models import Dataset
 from udata.core.organization.factories import OrganizationFactory
@@ -55,7 +57,8 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
         uri = url_for("api.organization_rdf", org=origin_org.id, _external=True)
 
         datasets = DatasetFactory.create_batch(3, organization=origin_org)
-        catalog = build_org_catalog(origin_org, datasets)
+        dataservices = DataserviceFactory.create_batch(3, organization=origin_org)
+        catalog = build_org_catalog(origin_org, datasets, dataservices)
 
         graph = catalog.graph
 
@@ -69,6 +72,7 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
         self.assertEqual(str(catalog.identifier), uri)
 
         self.assertEqual(len(list(catalog.objects(DCAT.dataset))), len(datasets))
+        self.assertEqual(len(list(catalog.objects(DCAT.dataservice))), len(dataservices))
 
         org = catalog.value(DCT.publisher)
         self.assertEqual(org.value(RDF.type).identifier, FOAF.Organization)
@@ -80,6 +84,8 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
         graph = catalog.graph
         graph_datasets = graph.subjects(RDF.type, DCAT.Dataset)
         self.assertEqual(len(list(graph_datasets)), len(datasets))
+        graph_dataservices = graph.subjects(RDF.type, DCAT.DataService)
+        self.assertEqual(len(list(graph_dataservices)), len(dataservices))
 
     def test_catalog_pagination(self):
         origin_org = OrganizationFactory()
@@ -102,11 +108,23 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
             page_size=page_size,
             _external=True,
         )
+        # First create a dataset and it's associated dataservice, which should be listed
+        # last, and thus on the second page.
+        extra_dataset = DatasetFactory.create(organization=origin_org)
+        _extra_dataservice = DataserviceFactory.create(
+            datasets=[extra_dataset], organization=origin_org
+        )
+
+        # Create `total` datasets that should be listed on the first page up to `page_size`
         DatasetFactory.create_batch(total, organization=origin_org)
+        # And all the dataservices with no datasets, which will all be listed on the first page.
+        # See DataserviceQuerySet.filter_by_dataset_pagination.
+        DataserviceFactory.create_batch(total, organization=origin_org)
 
         # First page
         datasets = Dataset.objects.paginate(1, page_size)
-        catalog = build_org_catalog(origin_org, datasets, format="json")
+        dataservices = Dataservice.objects.filter_by_dataset_pagination(datasets, 1)
+        catalog = build_org_catalog(origin_org, datasets, dataservices, format="json")
         graph = catalog.graph
 
         self.assertIsInstance(catalog, RdfResource)
@@ -115,9 +133,12 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
         self.assertIn(DCAT.Catalog, types)
         self.assertIn(HYDRA.Collection, types)
 
-        self.assertEqual(catalog.value(HYDRA.totalItems), Literal(total))
+        self.assertEqual(catalog.value(HYDRA.totalItems), Literal(total + 1))
 
         self.assertEqual(len(list(catalog.objects(DCAT.dataset))), page_size)
+        # All dataservices that are not linked to a dataset are listed in the first page.
+        # See DataserviceQuerySet.filter_by_dataset_pagination.
+        self.assertEqual(len(list(catalog.objects(DCAT.dataservice))), total)
 
         paginations = list(graph.subjects(RDF.type, HYDRA.PartialCollectionView))
         self.assertEqual(len(paginations), 1)
@@ -130,7 +151,8 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
 
         # Second page
         datasets = Dataset.objects.paginate(2, page_size)
-        catalog = build_org_catalog(origin_org, datasets, format="json")
+        dataservices = Dataservice.objects.filter_by_dataset_pagination(datasets, 2)
+        catalog = build_org_catalog(origin_org, datasets, dataservices, format="json")
         graph = catalog.graph
 
         self.assertIsInstance(catalog, RdfResource)
@@ -139,9 +161,12 @@ class OrganizationToRdfTest(DBTestMixin, TestCase):
         self.assertIn(DCAT.Catalog, types)
         self.assertIn(HYDRA.Collection, types)
 
-        self.assertEqual(catalog.value(HYDRA.totalItems), Literal(total))
+        self.assertEqual(catalog.value(HYDRA.totalItems), Literal(total + 1))
 
-        self.assertEqual(len(list(catalog.objects(DCAT.dataset))), 1)
+        # 5 datasets total, 3 on the first page, 2 on the second.
+        self.assertEqual(len(list(catalog.objects(DCAT.dataset))), 2)
+        # 1 extra_dataservice, listed on the same page as its associated extra_dataset.
+        self.assertEqual(len(list(catalog.objects(DCAT.dataservice))), 1)
 
         paginations = list(graph.subjects(RDF.type, HYDRA.PartialCollectionView))
         self.assertEqual(len(paginations), 1)